+++ ReadyCheckKick/ModEntry.cs
@@ -1,5 +1,6 @@
-using StardewModdingAPI;
+﻿using StardewModdingAPI;
 using StardewModdingAPI.Events;
+using weizinai.StardewValleyMod.PiCore.Extension;
 using weizinai.StardewValleyMod.PiCore.Integration.GenericModConfigMenu;
 using weizinai.StardewValleyMod.PiCore.Logging;
 using weizinai.StardewValleyMod.PiCore.Patcher;
@@ -27,9 +28,42 @@ internal class ModEntry : Mod
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
         this.AddGenericModConfigMenu(
-            new GenericModConfigMenuForReadyCheckKick(),
             () => ModConfig.Instance,
-            config => ModConfig.Instance = config
+            config => ModConfig.Instance = config,
+            configMenu => configMenu
+                // Show unready farmers
+                .AddSectionTitle(I18n.Config_ShowUnreadyFarmersTitle_Name)
+                .AddBoolOption(
+                    config => config.ShowInfoInReadyCheckDialogue,
+                    (config, value) => config.ShowInfoInReadyCheckDialogue = value,
+                    I18n.Config_ShowInfoInReadyCheckDialogue_Name
+                )
+                .AddBoolOption(
+                    config => config.ShowInfoInSaveGameMenu,
+                    (config, value) => config.ShowInfoInSaveGameMenu = value,
+                    I18n.Config_ShowInfoInSaveGameMenu_Name
+                )
+                // Kick unready farmers
+                .AddSectionTitle(I18n.Config_KickUnreadyFarmersTitle_Name)
+                .AddBoolOption(
+                    config => config.AutoKickUnreadyFarmers,
+                    (config, value) => config.AutoKickUnreadyFarmers = value,
+                    I18n.Config_AutoKickUnreadyFarmers_Name
+                )
+                .AddNumberOption(
+                    config => config.AutoKickUnreadyFarmersRatio,
+                    (config, value) => config.AutoKickUnreadyFarmersRatio = value,
+                    I18n.Config_AutoKickUnreadyFarmersRatio_Name,
+                    null,
+                    0f,
+                    1f,
+                    0.05f
+                )
+                .AddNumberOption(
+                    config => config.AutoKickUnreadyFarmersDelay,
+                    (config, value) => config.AutoKickUnreadyFarmersDelay = value,
+                    I18n.Config_AutoKickUnreadyFarmersDelay_Name
+                )
         );
     }
 }