+++ TestMod/ModEntry.cs
@@ -1,6 +1,7 @@
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
+using weizinai.StardewValleyMod.PiCore.Extension;
 using weizinai.StardewValleyMod.PiCore.Integration.GenericModConfigMenu;
 using weizinai.StardewValleyMod.PiCore.Patcher;
 using weizinai.StardewValleyMod.TestMod.Framework;
@@ -31,9 +32,39 @@ public class ModEntry : Mod
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
         this.AddGenericModConfigMenu(
-            new GenericModConfigMenuIntegrationForTestMod(),
             () => ModConfig.Instance,
-            value => ModConfig.Instance = value
+            value => ModConfig.Instance = value,
+            configMenu => configMenu
+                .AddSectionTitle(I18n.Config_CardChance_Name)
+                .AddBoolOption(
+                    config => config.CardChance.IsEnabled,
+                    (config, value) => config.CardChance.IsEnabled = value,
+                    I18n.Config_IsEnabled_Name
+                )
+                .AddNumberOption(
+                    config => config.CardChance.Value,
+                    (config, value) => config.CardChance.Value = value,
+                    I18n.Config_Value_Name
+                )
+                .AddSectionTitle(I18n.Config_WheelSpinSpeed_Name)
+                .AddBoolOption(
+                    config => config.WheelSpinSpeed.IsEnabled,
+                    (config, value) => config.WheelSpinSpeed.IsEnabled = value,
+                    I18n.Config_IsEnabled_Name
+                )
+                .AddNumberOption(
+                    config => config.WheelSpinSpeed.Value,
+                    (config, value) => config.WheelSpinSpeed.Value = value,
+                    I18n.Config_Value_Name,
+                    null,
+                    0,
+                    14
+                )
+                .AddBoolOption(
+                    config => config.ExtraSpeed,
+                    (config, value) => config.ExtraSpeed = value,
+                    I18n.Config_ExtraSpeed_Name
+                )
         );
     }